Fix source package name detection
The previous version assumed that the package path ended with the
package name. In cases where the package path contained invalid package
characters, like "-", the tools would generate invalid output.

The naive solution to this would be to use the `--package` flag to set
the name to something valid. However, this flag indicates that the
source and destination packages are different. This results in the
source package receiving an import alias to prevent conflicts. This
means that any package generated using `--package` that is rendered
within the file tree of `--source` also results in an invalid file.

This patch preserves all existing behaviors except that it now correctly
extracts the true package name rather than relying on the package path.